Calculator component split in other components. Callbacks moved
