+++ tutorlolv2_frontend/src/calculator_v2/mod.rs
@@ -1,4 +1,6 @@
 pub mod ability_selector;
+pub mod active_player;
+pub mod enemy_players;
 pub mod page;
 pub mod reducer;
